tests/wpt/web-platform-tests/fetch/metadata/sec-fetch-dest/fetch-preflight.tentative.https.sub.html

<!DOCTYPE html>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/fetch/metadata/resources/helper.js></script>
<script>
  // Site
  promise_test(t => {
    return fetch("https://{{hosts[][www]}}:{{ports[https][0]}}/fetch/metadata/resources/echo-as-json.py",
                 {
                   mode: "cors",
                   headers: { 'x-test': 'testing' }
                 })
        .then(r => r.json())
        .then(j => assert_header_dest_equals(j, "empty"));
  }, "Same-site fetch with preflight");

  promise_test(t => {
    return fetch("https://{{hosts[alt][www]}}:{{ports[https][0]}}/fetch/metadata/resources/echo-as-json.py",
                 {
                   mode: "cors",
                   headers: { 'x-test': 'testing' }
                 })
        .then(r => r.json())
        .then(j => assert_header_dest_equals(j, "empty"));
  }, "Cross-site fetch with preflight");
</script>
